refactor(tests): update setup_product_images method
Update the setup_product_images method in the test_views.py file under the tests directory of the products app. The method now uses the fixture 'products_with_images' instead of 'product' and 'image_no_save_file'. It also assigns the 6th product and the third image of that product as the target image.